Major code refactoring in views/job.py + Fix error in edit mode
Error was occuring when a job was created but not all parts of the model were created.
Now, even if a job has been created with only partial models, it will create the new model (e.g. create a new psf entry) and then continue on.

Refactoring now reduces considerably the redundancy in the code. Note that the return render(...) has to appear in the method that was originally called by the request -- which in itself pose the need to repeat ourselves... Something to think about.